refactor(kiloclaw): hoist instance-not-running sentinel to shared module
