+++ apps/web/src/lib/kiloclaw/types.ts
@@ -420,40 +420,16 @@ export type RestartMachineResponse = {
   error?: string;
 };
 
-/**
- * Sentinel returned by the polling endpoints (gateway/status,
- * controller-version, debug-status, morning-briefing/status) when the DO
- * state isn't `running`. The worker short-circuits before forwarding to
- * the Fly proxy so the proxy can't wake a stopped machine.
- *
- * Returned as HTTP 200 (not 503) so the frontend's high-frequency polling
- * doesn't generate log/Sentry noise for what is an expected steady state.
- * Frontends discriminate on `reason === 'instance_not_running'`.
- */
-export type InstanceNotRunningSentinel = {
-  ok: false;
-  reason: 'instance_not_running';
-  /** Current DO status. Frontend can use this to render a richer label
-   *  (e.g. "Instance is starting…" vs "Instance is stopped"). */
-  status:
-    | 'provisioned'
-    | 'starting'
-    | 'restarting'
-    | 'recovering'
-    | 'stopped'
-    | 'destroying'
-    | 'restoring'
-    | null;
-};
-
-export function isInstanceNotRunningSentinel(value: unknown): value is InstanceNotRunningSentinel {
-  return (
-    typeof value === 'object' &&
-    value !== null &&
-    'reason' in value &&
-    (value as { reason?: unknown }).reason === 'instance_not_running'
-  );
-}
+// The sentinel type + predicate live in a dep-free module so apps/mobile
+// can pick them up via tsconfig path alias (mirroring the existing
+// `@/lib/images-schema` cross-package import pattern). Re-exported here so
+// existing imports from `@/lib/kiloclaw/types` keep working.
+import {
+  isInstanceNotRunningSentinel,
+  type InstanceNotRunningSentinel,
+} from './instance-not-running-sentinel';
+export { isInstanceNotRunningSentinel };
+export type { InstanceNotRunningSentinel };
 
 /**
  * Narrowing helper: returns the OK-shape variant of a polling-endpoint
